chore: multi-file structure
