+++ packages/console/src/components/Breadcrumbs/validators/index.ts
@@ -1,0 +1,4 @@
+export const breadcrumbDefaultvalidator = (
+  urlPathId: string,
+  thisPathId: string,
+) => urlPathId === thisPathId;
